S7: changed byteLength and blockNumber from short to int
(Intentionally trying this via the web ui) ...

+++ plc4j/protocols/s7/src/main/java/org/apache/plc4x/java/s7/netty/model/params/items/S7AnyVarParameterItem.java
@@ -44,11 +44,11 @@ public class S7AnyVarParameterItem implements VarParameterItem {
     private final MemoryArea memoryArea;
     private final TransportSize dataType;
     private final int numElements;
-    private final short dataBlockNumber;
-    private final short byteOffset;
+    private final int dataBlockNumber;
+    private final int byteOffset;
     private final byte bitOffset;
 
-    public S7AnyVarParameterItem(SpecificationType specificationType, MemoryArea memoryArea, TransportSize dataType, int numElements, short dataBlockNumber, short byteOffset, byte bitOffset) {
+    public S7AnyVarParameterItem(SpecificationType specificationType, MemoryArea memoryArea, TransportSize dataType, int numElements, int dataBlockNumber, int byteOffset, byte bitOffset) {
         this.specificationType = specificationType;
         this.memoryArea = memoryArea;
         this.dataType = dataType;
@@ -79,11 +79,11 @@ public class S7AnyVarParameterItem implements VarParameterItem {
         return numElements;
     }
 
-    public short getDataBlockNumber() {
+    public int getDataBlockNumber() {
         return dataBlockNumber;
     }
 
-    public short getByteOffset() {
+    public int getByteOffset() {
         return byteOffset;
     }
